+++ examples/example_server_endpoint.py
@@ -7,7 +7,6 @@ from common.sim_env import BOX_POSE, make_sim_env
 
 import neuracore as nc
 from neuracore import EndpointError
-from neuracore.core.nc_types import DataType
 
 ENDPOINT_NAME = "MyExampleEndpoint"
 
@@ -60,11 +59,16 @@ def main():
                     nc.log_rgb(key, value)
             idx_in_horizon = i % horizon
             if idx_in_horizon == 0:
-                prediction = policy.predict()
-                action = prediction.outputs[DataType.JOINT_TARGET_POSITIONS]
-                horizon = action.shape[0]
+                predicted_sync_points = policy.predict()
+                joint_target_positions = [
+                    sp.joint_target_positions for sp in predicted_sync_points
+                ]
+                actions = [
+                    jtp.numpy() for jtp in joint_target_positions if jtp is not None
+                ]
+                horizon = len(actions)
 
-            a = action[idx_in_horizon]
+            a = actions[idx_in_horizon]
             ts = env.step(a)
             episode_max = max(episode_max, ts.reward)