+++ tests/unit/ml/algorithms/test_act.py
@@ -50,7 +50,9 @@ def model_init_description() -> ModelInitDescription:
             mean=np.zeros(JOINT_POSITION_DIM, dtype=float),
             std=np.ones(JOINT_POSITION_DIM, dtype=float),
         ),
-        max_num_rgb_images=CAMS,
+        rgb_images=DataItemStats(
+            max_len=CAMS,
+        ),
     )
     return ModelInitDescription(
         dataset_description=dataset_description,